+++ griptape/drivers/audio_transcription/dummy_audio_transcription_driver.py
@@ -5,7 +5,7 @@ from typing import TYPE_CHECKING, Optional
 from attrs import define, field
 
 from griptape.drivers import BaseAudioTranscriptionDriver
-from griptape.exceptions import DummyException
+from griptape.exceptions import DummyError
 
 if TYPE_CHECKING:
     from griptape.artifacts import AudioArtifact, TextArtifact
@@ -16,4 +16,4 @@ class DummyAudioTranscriptionDriver(BaseAudioTranscriptionDriver):
     model: str = field(init=False)
 
     def try_run(self, audio: AudioArtifact, prompts: Optional[list] = None) -> TextArtifact:
-        raise DummyException(__class__.__name__, "try_transcription")
+        raise DummyError(__class__.__name__, "try_transcription")