+++ griptape/drivers/text_to_speech/dummy_text_to_speech_driver.py
@@ -5,7 +5,7 @@ from typing import TYPE_CHECKING
 from attrs import define, field
 
 from griptape.drivers import BaseTextToSpeechDriver
-from griptape.exceptions import DummyException
+from griptape.exceptions import DummyError
 
 if TYPE_CHECKING:
     from griptape.artifacts.audio_artifact import AudioArtifact
@@ -16,4 +16,4 @@ class DummyTextToSpeechDriver(BaseTextToSpeechDriver):
     model: None = field(init=False, default=None, kw_only=True)
 
     def try_text_to_audio(self, prompts: list[str]) -> AudioArtifact:
-        raise DummyException(__class__.__name__, "try_text_to_audio")
+        raise DummyError(__class__.__name__, "try_text_to_audio")