+++ tests/unit/drivers/embedding/test_dummy_embedding_driver.py
@@ -1,7 +1,7 @@
 import pytest
 
 from griptape.drivers import DummyEmbeddingDriver
-from griptape.exceptions import DummyException
+from griptape.exceptions import DummyError
 
 
 class TestDummyEmbeddingDriver:
@@ -13,5 +13,5 @@ class TestDummyEmbeddingDriver:
         assert embedding_driver
 
     def test_try_embed_chunk(self, embedding_driver):
-        with pytest.raises(DummyException):
+        with pytest.raises(DummyError):
             embedding_driver.try_embed_chunk("prompt-stack")